requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.3.3
openai==1.63.2
chromadb==0.6.3
streamlit==1.42.0
pypdf==5.3.0
python-dotenv==1.0.1
tiktoken==0.9.0
beautifulsoup4==4.12.3
requests==2.32.3
pydantic>=2.0.0